requirements.txt
fastapi==0.116.1
openai==1.96.1
pandas==2.3.1
pdfplumber==0.11.7
pinecone==7.3.0
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn[standard]
python-multipart